// blazor-grid-components/scripts/interfaces.ts
export interface VirtualInfo {
    data?: boolean;
    event?: string;
    block?: number;
    page?: number;
    currentPage?: number;
    direction?: string;
    blockIndexes?: number[];
    columnIndexes?: number[];
    columnBlocks?: number[];
    loadSelf?: boolean;
    loadNext?: boolean;
    nextInfo?: { page?: number };
    sentinelInfo?: SentinelType;
    offsets?: Offsets;
    startIndex?: number;
    endIndex?: number;
}

/**
 * @hidden
 */
export interface InterSection {
    container?: HTMLElement;
    pageHeight?: number;
    debounceEvent?: boolean;
    axes?: string[];
    virtualScrollHandler?: Function;
    totalItems?: number;
    isWheelScroll?: boolean;
    overscanCount?: number;
    rowHeight?: number;
    maskRowEnabled?: boolean;
    height?: number;
    viewInfo?: VirtualInfo;
    pageSize?: number;
    previousTop?: number;
    previousLeft?: number;
}